DapperORM/ContactRepository.cs
using System;
using System.Collections.Generic;
using System.Data;
using System.Data.SqlClient;
using System.Text;
using System.Configuration;
using Dapper;
using System.Linq;

namespace DapperORM
{
    public class ContactRepository : IContactRepository 
    {
        IDbConnection db;
       public ContactRepository()
        {
            db = new SqlConnection("Data Source=.;Initial Catalog=Test2;Integrated Security=True");
        }
        public Contact Add(Contact contact)
        {
            throw new NotImplementedException();
        }

        public Contact Find(int id)
        {
            throw new NotImplementedException();
        }

        public List<Contact> GetAll()
        {
            return this.db.Query<Contact>("Select * from Contacts").ToList();
        }

        public void Remove(int id)
        {
            throw new NotImplementedException();
        }

        public void Save(Contact contact)
        {
            throw new NotImplementedException();
        }

        public Contact Update(Contact contact)
        {
            throw new NotImplementedException();
        }
    }
}
